Moved database files to new package
Made copies of lists in GroupDataAccess.java to avoid errors and directly modifying lists
